+++ include/vec/ranges.h
@@ -5,8 +5,12 @@
 
 Vec vec_slice (Vec v, long start, long end);
 Vec vec_subvec (Vec v, long start, long length);
+
 Vec vec_prefix (Vec v, long length);
 Vec vec_suffix (Vec v, long length);
+
+Vec vec_copy (Vec v);
+Vec vec_duplicate (Vec v);