+++ include/vec/types.h
@@ -37,6 +37,9 @@ struct VecNamespace {
 	Vec (*prefix) (Vec v, long length);
 	Vec (*suffix) (Vec v, long length);
 
+	Vec (*copy) (Vec v);
+	Vec (*duplicate) (Vec v);
+
 	VecNamespace (*push) (Vec v, void *e);
 	void *(*pop) (Vec v);