+++ lib/vec.c
@@ -28,6 +28,9 @@ struct VecNamespace vec_struct = {
 	.prefix = &vec_prefix,
 	.suffix = &vec_suffix,
 
+	.copy = &vec_copy,
+	.duplicate = &vec_duplicate,
+
 	.push = &vec_push,
 	.pop = &vec_pop,